Fixed some simulation issues.
We had issues with the fact that simulations of options were not correct according to their theoretical value. For example, when comparing a simulation of a european call option, the result did not correspond with the Black Scholes formula. The soulution for now was to work with year-dates instead of days for now.
This is a temporary fix.